// frontend/src/components/Vault/block-editor/page-editor/types.ts
export type PageOption = string | { name: string; color?: string; group?: string;[key: string]: unknown };
export interface PagePropertyConfig extends Record<string, unknown> {
  id?: string;
  options?: PageOption[];
  description?: string;
  format?: FieldFormat | null;
}
export interface PageProperty extends PagePropertyConfig {
  name: string;
  type: string;
  config?: PagePropertyConfig;
  relation_database_id?: string;
  file_mode?: string;
  storage_folder?: string;
  name_pattern?: string;
}